split runners into their own file, start worrying about feature cleanliness
