+++ .github/workflows/run-tests.yml
@@ -50,8 +50,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        runner: [orig, single]
-    name: Runner ${{ matrix.runner }}
+        run: [orig, single]
+    name: Runner ${{ matrix.run }}
     runs-on: ubuntu-latest
     steps:
     - uses: actions/cache@v3
@@ -69,9 +69,38 @@ jobs:
     - name: Set up Rust
       run: |
         rustup default stable
-        rustup component add rustfmt
 
     - uses: actions/checkout@v3
     - name: Deploy events
       run: |
-        cargo run --release --bin mass-events --features sys_nop,run_${{ matrix.runner }}
+        cargo run --release --bin mass-events --features sys_nop,run_${{ matrix.run }}
+
+
+  exhaustive-features:
+    strategy:
+      fail-fast: false
+      matrix:
+        run: [orig, single]  # TODO: std_multi, rayon
+        sys: [nop, cli, gui_softbuffer]
+    name: Checking run_${{ matrix.run }} and sys_${{ matrix.sys }}
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/cache@v3
+      with:
+        path: ~/.rustup/
+        key: rustup-ubuntu-stable
+    - uses: actions/cache@v3
+      with:
+        path: ~/.cargo/registry
+        key: cargo-cache
+    - uses: actions/cache@v3
+      with:
+        path: ./target
+        key: target-dir-ubuntu-stable
+    - name: Set up Rust
+      run: |
+        rustup default stable
+
+    - name: Check feature compatibility
+      run: |
+        cargo check --features sys_${{ matrix.sys }},run_${{ matrix.run }} --all-targets
